Huge changes! The parser now works in the IO monad.  This took a bunch of work and involved many changes to types in TypeAG.hs and in SolarmanTriplestore.hs This is to facilitate the proper way of doing IO in Haskell with our triple store.  Now, the TripleStore typeclass returns things in the IO monad, so that we can do SPARQL queries and whatnot.
There's a problem, though: the parser code itself in AGParser2.hs
is all pure, so actually displaying the results is tricky.
Right now, a simple hack using unsafePerformIO (the very
thing I wanted to avoid in the first place...) is used
to actually display the values.

This branch is experimental.  If we decide to just use
unsafePerformIO directly in our getts_* functions,
then this branch isn't needed.  It's not proper, but it'll work,
and Dr. Frost will probably be okay with that.  But
in the event we want to do things the proper way,
we've got to figure out how to get the parser to format
things in the IO monad (it uses show right now, which
won't work since that's a pure function).

So this code works, but still isn't a complete solution
to the unsafePerformIO problem